vinay: 18 June 2014 -These were the changes made last time — Merged changes in the updated GaborRFMap to the previous version of CRSMap (when it was at the same stage as the earlier GaborRF); defined GaborRF as a parallel branch — Then merged this updated GaborRFMap with the current CRSMap branch and this now becomes the working branch
For this commit - Got this updated CRSMap to work on Mavericks for the first time
Initially this was giving linker errors during ‘Build’, even after all the modifications were done. I had to manually add ‘CRSUtilities.h’ and ‘CRSUtilities.m’ to the project for it to ‘Build’ without errors. These files weren’t there in the earlier CRSMap or the old GaborRFMap